he eliminado la clase CarsController y he copiado su metodo select a la clase CarsControllers
