+++ src/AppForSEII2526.API/Controllers/CarsControllers.cs
@@ -58,8 +58,47 @@ namespace AppForSEII2526.API.Controllers
         }
 
 
+        [HttpGet]
+        [Route("[action]")]
+        [ProducesResponseType(typeof(IEnumerable<CocheParaReviewDTO>), (int)HttpStatusCode.OK)]
+        [ProducesResponseType((int)HttpStatusCode.BadRequest)]
+        public async Task<ActionResult<IEnumerable<CocheParaReviewDTO>>> GetCars2(string? FuelType = null)
+        {
+            try
+            {
+                var query = _context.Cars.AsQueryable();
 
+                // Aplicar filtro SOLO por color (flujo alternativo)
+                if (!string.IsNullOrEmpty(FuelType))
+                    query = query.Where(c => c.FuelType.Contains(FuelType));
 
+                var cars = await query
+                    .Select(c => new CocheParaReviewDTO(
+                        c.Id,
+                        c.Model.Name,
+                        c.Color,
+                        c.CarClass,
+                        c.Manufacturer,
+                        c.FuelType
+                    ))
+                    .ToListAsync();
 
+                if (cars == null || !cars.Any())
+                    return NotFound("No se encontraron coches con ese tipo de combustible.");
+
+                return Ok(cars);
+            }
+            catch (Exception ex)
+            {
+                _logger.LogError($"{DateTime.Now}: Error en GetCars() - {ex.Message}");
+                return BadRequest("Error al obtener los coches disponibles.");
+            }
+        }
     }
+
+
+
+
+
 }
+
